Remove all TkOnly dependencies from SLHCUpgradeSimulation package
