+++ L1Trigger/TrackTrigger/test/SLHC_PGUN_TkOnly_TILTED_81X.py
@@ -126,8 +126,11 @@ for path in process.paths:
 	getattr(process,path)._seq = process.generator * getattr(process,path)._seq
 	
 # Automatic addition of the customisation function from SLHCUpgradeSimulations.Configuration.combinedCustoms
-from SLHCUpgradeSimulations.Configuration.combinedCustoms import cust_2023tilted 
+
+from L1Trigger.TrackTrigger.TkOnlyDigi_cff import TkOnlyDigi
+from SLHCUpgradeSimulations.Configuration.combinedCustoms import cust_2023tilted
 
 process = cust_2023tilted(process)
+process = TkOnlyDigi(process)
 # End of customisation functions
