+++ L1Trigger/TrackTrigger/test/SLHC_PU_TkOnly_FLAT_81X.py
@@ -129,9 +129,11 @@ for path in process.paths:
 
 
 # Automatic addition of the customisation function from SLHCUpgradeSimulations.Configuration.combinedCustoms
-from SLHCUpgradeSimulations.Configuration.combinedCustoms import cust_2023LReco 
+from L1Trigger.TrackTrigger.TkOnlyDigi_cff import TkOnlyDigi
+from SLHCUpgradeSimulations.Configuration.combinedCustoms import cust_2023flat
 
-process = cust_2023LReco(process)
+process = cust_2023flat(process)
+process = TkOnlyDigi(process)
 # End of customisation functions
